+++ CHANGELOG.md
@@ -10,6 +10,8 @@ Note: In this file, do not use the hard wrap in the middle of a sentence for com
 
 ## [Unreleased]
 
+- Track location that `AssertUnmoved` being first pinned-mutably accessed.
+
 ## [0.1.4] - 2022-02-06
 
 - Use `#[track_caller]`.